% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positional_usage.R
\name{compute_diff_pssm}
\alias{compute_diff_pssm}
\title{Compute differential position specific scoring matrix (PSSM) between two groups (applicable to peptides of same lengths)}
\usage{
compute_diff_pssm(pos_peptides, neg_peptides)
}
\arguments{
\item{pos_peptides}{list of peptides analyzed for enrichment}

\item{neg_peptides}{list of peptides analyzed for depletion}
}
\value{
differnetial pssm matrix
}
\description{
Compute differential position specific scoring matrix (PSSM) between two groups (applicable to peptides of same lengths)
}
\examples{
diff_pssm_mtx <- compute_diff_pssm(pos_peptides, neg_peptides)

}
