% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{compute_ngrams}
\alias{compute_ngrams}
\title{Compute enrichment or depletion of n-grams between two groups}
\usage{
compute_ngrams(pos_peptides, neg_peptides, ngram_lengths = c(2, 3, 4, 5))
}
\arguments{
\item{pos_peptides}{list of peptides analyzed for enrichment}

\item{neg_peptides}{list of peptides analyzed for depletion}

\item{ngram_lengths}{list of n to compute for n-grams (e.g. 2 for n-gram)}
}
\value{
data table summarizing number of n-grams appeared in positive and negative peptide lists
and ratio of counts normalized by the total number of positive and negative peptides respectively
}
\description{
Compute enrichment or depletion of n-grams between two groups
}
\examples{
 ngram_df <- compute_ngrams(pos_peptides, neg_peptides, ngram_lengths = c(2, 3, 4))

}
