% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{compute_pairwise_alignment}
\alias{compute_pairwise_alignment}
\title{Compute pairwise alignment between peptides}
\usage{
compute_pairwise_alignment(
  peptides,
  substitutionMatrix = "BLOSUM62",
  alignType = "global",
  gapOpening = 10,
  gapExtension = 4
)
}
\arguments{
\item{substitutionMatrix}{substitution matrix to compute pairwise alignment score,
examples include 'BLOSUM62', 'BLOSUM50' or can input matrix of interest e.g. \code{data('BLOSUM62')}}

\item{alignType}{type of alignment, options include "global", "local", "overlap",
"global-local", and "local-global"}

\item{gapOpening}{cost for opening a gap in the alignment.}

\item{gapExtension}{the incremental cost incurred along the length of the gap in the alignment.}
}
\value{
alignment matrix
}
\description{
Compute pairwise alignment between peptides
}
\examples{
adjacency_matrix <- compute_pairwise_distance(peptides = c(pos_peptides, neg_peptides))
adjacency_matrix_positive <- compute_pairwise_alignment(peptides = pos_peptides)
adjacency_matrix_negative <- compute_pairwise_alignment(peptides = neg_peptides)

}
