% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{compute_pairwise_distance}
\alias{compute_pairwise_distance}
\title{Compute distance among peptides}
\usage{
compute_pairwise_distance(peptides, method = "hamming")
}
\arguments{
\item{method}{Method for distance calculation, default is 'hamming'. Options include
"osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", soundex".
See \code{'stringdist-metric'} for more details.}
}
\value{
distance matrix
}
\description{
Compute distance among peptides
}
\examples{
distance_matrix <- compute_pairwise_distance(peptides = c(pos_peptides, neg_peptides))

}
