% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positional_kmer.R
\name{compute_positional_kmers}
\alias{compute_positional_kmers}
\title{Compute enrichment or depletion of positional k-mers between two groups (applicable to peptides of same lengths)}
\usage{
compute_positional_kmers(pos_peptides, neg_peptides)
}
\value{
data table summarizing number of positional k-mer motifs appeared in postiive and negative peptide lists
and ratio of counts normalized by the total number of positive and negative peptides respectively
}
\description{
Compute enrichment or depletion of positional k-mers between two groups (applicable to peptides of same lengths)
}
\examples{
kmer_df <- compute_positional_kmers(pos_peptides, neg_peptides)

}
