% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffSeqPatterns.R
\docType{package}
\name{diffSeqPatterns}
\alias{diffSeqPatterns}
\title{diffSeqPatterns: R package for analyzing differential peptide sequence patterns between two groups}
\arguments{
\item{peptides}{list of peptides}

\item{pos_peptides}{list of peptides analyzed for enrichment}

\item{neg_peptides}{list of peptides analyzed for depletion}
}
\description{
The diffSeqPatterns package contains functions to analyze and visualize
enrichment or depletion of sequence patterns between two groups,
such as amino acid usage, positional k-mer motifs, n-grams, and inter-sequence distances.
}
\details{
It addresses the following needs:
\itemize{ \item Compute differential sequence patterns between two groups
\item Effectively visualize patterns;
\item Can be applied to any set of two peptide groups;
}
}
\section{Section}{

Common arguments to the functions.
}

