% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DMF5_antigen_table}
\alias{DMF5_antigen_table}
\title{Table of 55 DMF5 antigens and 200 background peptides annotated for their T cell specificity}
\format{
data table with peptides at their contact positions and their specificity against DMF5 T cells
}
\usage{
DMF5_antigen_table
}
\description{
This is table used for color annotations
}
\keyword{datasets}
