% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DMF5_neg_peptides}
\alias{DMF5_neg_peptides}
\title{Negative peptide list from DMF5 selection}
\format{
A list of 200 peptides
}
\usage{
DMF5_neg_peptides
}
\description{
This is list of peptide sequences at contact positions annotated 'Negative'
from DMF5 T cell screening, analyzed for depletion.
Used for demonstrations in vignette and application note.
}
\keyword{datasets}
