% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DMF5_pos_peptides}
\alias{DMF5_pos_peptides}
\title{Positive peptide list from DMF5 selection}
\format{
A list of 55 peptides
}
\usage{
DMF5_pos_peptides
}
\description{
This is list of peptide sequences at contact positions annotated 'Positive'
against DMF5 T cell screening, analyzed for enrichment
}
\keyword{datasets}
