% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iedbdata}
\alias{iedbdata}
\title{This is selected peptide data from IEDB database}
\format{
A data frame with 1733 rows and 3 variables:
\describe{
  \item{Peptide}{Original 9aa peptide sequence}
  \item{Immunogenicity}{Positive or Negative denoting whether the peptide has been
  characterized to trigger immune response by T cell assays}
  \item{ContactPositions}{contact residues (positions 3-8) of the original peptide seqquence}
}
}
\usage{
iedbdata
}
\description{
This is selected peptide data from IEDB database
}
\keyword{datasets}
