% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neg_peptides}
\alias{neg_peptides}
\title{Negative peptide list}
\format{
A list of 939 peptides
}
\usage{
neg_peptides
}
\description{
This is list of peptide sequences annotated 'Negative' from T cell assays, analyzed for depletion
}
\keyword{datasets}
