% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positional_kmer.R
\name{plot_bar_posiitonal_kmers}
\alias{plot_bar_posiitonal_kmers}
\title{Visualize top enriched and depleted k-mer motifs in barplot}
\usage{
plot_bar_posiitonal_kmers(
  positional_kmer_data,
  top_n = 10,
  ratio_threshold = 4,
  n_threshold = 6
)
}
\arguments{
\item{positional_kmer_data}{output matrix from \code{compute_positional_kmers}}

\item{top_n}{number of top k-mer motifs from positive and negative (respectively) to visualize}

\item{ratio_threshold}{ratio threshold to label k-mer motifs}

\item{n_threshold}{for k-mer motifs that are only present in either positive or negative list,
threshold to label k-mer motifs, e.g. n_threshold = 4, only label motif when there are >= 4 peptides
containing the motif in either positive or negative peptide lists.}
}
\value{
barplot
}
\description{
Visualize top enriched and depleted k-mer motifs in barplot
}
\examples{
positional_kmer_data <- compute_positional_kmers(pos_peptides, neg_peptides)
plot_bar_posiitonal_kmers(positional_kmer_data, top_n = 20)

}
