% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{plot_bar_top_ngrams}
\alias{plot_bar_top_ngrams}
\title{Visualize top enriched and depleted n-grams in barplot}
\usage{
plot_bar_top_ngrams(
  ngram_table,
  top_n = 10,
  ratio_threshold = 4,
  n_threshold = 6
)
}
\arguments{
\item{ngram_table}{output matrix from \code{compute_ngrams}}

\item{top_n}{number of top n-grams from positive and negative (respectively) to visualize}

\item{ratio_threshold}{ratio threshold to label n-grams}

\item{n_threshold}{for n-grams that are only present in either positive or negative list,
threshold to label n-grams, e.g. n_threshold = 4, only label motif when there are >= 4 peptides
containing the motif in either positive or negative peptide lists.}
}
\value{
barplot
}
\description{
Visualize top enriched and depleted n-grams in barplot
}
\examples{
ngram_df <- compute_ngrams(pos_peptides, neg_peptides, ngram_lengths = c(2, 3, 4))
p <- plot_bar_top_ngrams(ngram_df, top_n = 20)

}
