% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{plot_heatmap_alignment_mtx}
\alias{plot_heatmap_alignment_mtx}
\title{Plot heatmap from pairwise alignment matrix}
\usage{
plot_heatmap_alignment_mtx(adjacency_matrix, col_data)
}
\arguments{
\item{adjacency_matrix}{output from \code{compute_pairwise_alignment}}

\item{col_data}{dataframe that includes columns for color annotations and rownames of
peptides in adjacency matrix}
}
\value{
heatmap colored by distance between peptide sequences
}
\description{
Plot heatmap from pairwise alignment matrix
}
\examples{
adjacency_matrix <- compute_pairwise_distance(peptides = c(pos_peptides[1:50], neg_peptides[1:60]))
adjacency_matrix_positive <- compute_pairwise_distance(peptides = pos_peptides[1:50])

iedbdata <- read.csv(file = 'peptide_data.csv')
col_data <- iedbdata \%>\% filter(ContactPositions \%in\% colnames(adjacency_matrix)) \%>\%
  select(ContactPositions, Immunogenicity) \%>\% distinct() \%>\% column_to_rownames(var =  'ContactPositions')
p1 <- plot_heatmap_alignment_mtx(adjacency_matrix, col_data)
p2 <- plot_heatmap_alignment_mtx(adjacency_matrix_positive, col_data)

}
