% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{plot_heatmap_distance_mtx}
\alias{plot_heatmap_distance_mtx}
\title{Plot heatmap from distance adjacency matrix}
\usage{
plot_heatmap_distance_mtx(distance_mtx, col_data)
}
\arguments{
\item{distance_mtx}{output from \code{compute_pairwise_distance}}

\item{col_data}{dataframe that includes columns for color annotations and rownames of
peptides in adjacency matrix}
}
\value{
heatmap colored by distance between peptide sequences
}
\description{
Plot heatmap from distance adjacency matrix
}
\examples{
distance_mtx <- compute_pairwise_distance(peptides = c(pos_peptides[1:50], neg_peptides[1:60]))
col_data <- iedbdata \%>\% filter(ContactPositions \%in\% colnames(adjacency_matrix)) \%>\%
  select(ContactPositions, Immunogenicity) \%>\% distinct() \%>\%
  column_to_rownames(var =  'ContactPositions')
p <- plot_heatmap_distance_mtx(distance_mtx, col_data)

}
