% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{plot_network_alignment_mtx}
\alias{plot_network_alignment_mtx}
\title{Plot network graph from pairwise alignment matrix}
\usage{
plot_network_alignment_mtx(
  adjacency_mtx,
  data = NULL,
  peptide_id_col = "ContactPositions",
  color_col = "Immunogenicity",
  alignment_threshold = 0,
  vertex.label.degree = 0,
  edge.weight = 0.2,
  vertex.size = 5,
  vertex.label.cex = 0.8,
  label_vertex = FALSE
)
}
\arguments{
\item{data}{data to denote node colors - minimum columns are peptide and color_id}

\item{peptide_id_col}{column name for peptide_id e.g. Peptide}

\item{color_col}{column name for color annotation e.g. Immunogenicity}

\item{vertex.label.degree}{position of node label}

\item{edge.weight}{distance*edge.weight to represent thickness of edges in network plot}

\item{vertex.size}{node size}

\item{vertex.label.cex}{node font size}

\item{label_vertex}{TRUE/FALSE to print sequence for each node}

\item{adjacency_matrix}{output from \code{compute_pairwise_alignment}}
}
\value{
Network graph showing distance between peptide sequences
}
\description{
Plot network graph from pairwise alignment matrix
}
\examples{
iedbdata <- read.csv(file = 'peptide_data.csv')
adjacency_matrix <- compute_pairwise_distance(peptides = c(pos_peptides, neg_peptides))
p1 <- plot_network_alignment_mtx(adjacency_matrix, data = NULL)
p2 <- plot_network_alignment_mtx(adjacency_matrix, data = iedbdata,
                           peptide_id_col = 'ContactPositions',  color_col = 'Immunogenicity',
                           vertex.label.degree = 0,  edge.weight = 0.2,
                           vertex.size= 5,  vertex.label.cex = 0.8, label_vertex=TRUE)
p3 <- plot_network_alignment_mtx(adjacency_matrix, data = iedbdata,
                           peptide_id_col = 'ContactPositions',  color_col = 'Immunogenicity',
                           edge.weight = 0.2,  vertex.size= 5,   label_vertex=FALSE)

}
