% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positional_kmer.R
\name{plot_point_positional_kmers}
\alias{plot_point_positional_kmers}
\title{Visualize enrichment or depletion score of k-mer motifs in scatterplot}
\usage{
plot_point_positional_kmers(
  positional_kmer_data,
  ratio_threshold = 4,
  n_threshold = 6
)
}
\arguments{
\item{positional_kmer_data}{output matrix from \code{compute_positional_kmers}}

\item{ratio_threshold}{ratio threshold to label k-mer motifs}

\item{n_threshold}{for k-mer motifs that are only present in either positive or negative list,
threshold to label k-mer motifs, e.g. n_threshold = 4, only label motif when there are >= 4 peptides
containing the motif in either positive or negative peptide lists.}
}
\value{
scatterplot
}
\description{
Visualize enrichment or depletion score of k-mer motifs in scatterplot
}
\examples{
positional_kmer_data <- compute_positional_kmers(pos_peptides, neg_peptides)
p <- plot_point_positional_kmers(positional_kmer_data)


}
