% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pos_peptides}
\alias{pos_peptides}
\title{Positive peptide list}
\format{
A list of 794 peptides
}
\usage{
pos_peptides
}
\description{
This is list of peptide sequences annotated 'Positive' from T cell assays, analyzed for enrichment
}
\keyword{datasets}
