% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{plot_point_ngrams}
\alias{plot_point_ngrams}
\title{Visualize enrichment or depletion score of n-grams in scatterplot}
\usage{
plot_point_ngrams(ngram_table, ratio_threshold = 4, n_threshold = 6)
}
\arguments{
\item{ngram_table}{output matrix from \code{compute_ngrams}}
}
\value{
scatterplot
}
\description{
Visualize enrichment or depletion score of n-grams in scatterplot
}
\examples{
ngram_df <- compute_ngrams(pos_peptides, neg_peptides, ngram_lengths = c(2, 3, 4))
p <- plot_point_ngrams(ngram_df)

}
