% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positional_usage.R
\name{plot_seqlogo_diff_pssm}
\alias{plot_seqlogo_diff_pssm}
\title{Plot differential PSSM into sequence logos}
\usage{
plot_seqlogo_diff_pssm(pos_peptides, neg_peptides)
}
\arguments{
\item{pos_peptides}{list of peptides analyzed for enrichment}

\item{neg_peptides}{list of peptides analyzed for depletion}
}
\value{
sequence logo plot
}
\description{
Plot differential PSSM into sequence logos
}
\examples{
diff_pssm_mtx <- compute_diff_pssm(pos_peptides, neg_peptides)
p <- plot_seqlogo_diff_pssm(pos_peptides, neg_peptides)

}
